fix NavigationHost
